+++ Src/Obfy.Core/Services/ObfuscationService.cs
@@ -3,6 +3,7 @@ using System.Text.Json;
 using Microsoft.Extensions.Logging;
 using Obfy.Core.Models;
 using Obfy.Core.Pipeline;
+using Obfy.Core.Utilities;
 
 namespace Obfy.Core.Services;
 
@@ -79,8 +80,7 @@ public class ObfuscationService : IObfuscationService
 
             context.InputPath = inputPath;
             context.OutputPath = target.EffectiveOutputPath;
-
-
+            RuntimeProfileGating.Apply(settings, context);
         }
         catch (Exception ex)
         {